+++ README.md
@@ -92,10 +92,11 @@
 
 ### 1.5 투자 상품 등록
 
-- 사용자는 투자 상품을 등록합니다.
+- 사용자는 투자 상품을 등록할 수 있다.
     - 사용자는 여러개의 투자 상품을 가질 수 있습니다.
     - 투자 상품에는 예금, 적금 등이 존재합니다.
-- 등록된 투자 상품에는 사용자의 식별자, 투자 상품이 저장됩니다.
+- 투자 상품에는 식별자, 사용자 식별자, 투자 정보 등이 저장됩니다.
+- 사용자는 여러개의 투자 상품을 가질 수 있다.
 
 ## 2. 비기능 요구사항
