--- src/test/java/co/invest72/investment/application/AddInvestmentResponse.java
@@ -1,7 +1,0 @@
-package co.invest72.investment.application;
-
-import co.invest72.investment.domain.Investment;
-
-public record AddInvestmentResponse(Long id, String uid, Investment investment) {
-
-}
